Change the url param into body

+++ src/main/java/com/zetcco/jobscoutserver/services/BlogPostService.java
@@ -1,5 +1,6 @@
 package com.zetcco.jobscoutserver.services;
 
+import java.nio.file.AccessDeniedException;
 import java.util.Date;
 import java.util.List;
 
@@ -9,7 +10,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.zetcco.jobscoutserver.domain.BlogPost;
+import com.zetcco.jobscoutserver.domain.support.User;
 import com.zetcco.jobscoutserver.repositories.BlogPostRepository;
+import com.zetcco.jobscoutserver.repositories.UserRepository;
+import com.zetcco.jobscoutserver.services.mappers.UserMapper;
+import com.zetcco.jobscoutserver.services.support.BlogPostDTO;
+import com.zetcco.jobscoutserver.services.support.NotFoundException;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class BlogPostService {
@@ -20,9 +28,15 @@ public class BlogPostService {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    @Autowired
+    private UserMapper userMapper;
+
     public List<BlogPost> getBlogPosts(int pageNo, int pageSize) {
         Pageable page = PageRequest.of(pageNo, pageSize);
-        return blogPostRepository.findAll(page).getContent();
+        blogPostRepository.findAll(page).getContent();
     }
 
     public BlogPost getBlogPost(Long id) {
@@ -33,39 +47,49 @@ public class BlogPostService {
         return blogPost;
     }
 
-    public BlogPost updateBlogPost(Long id, String content, Date date) {
-        BlogPost blogPost = blogPostRepository.findById(id).orElseThrow();
+    @Transactional
+    public BlogPostDTO saveBlogPost(BlogPostDTO blogPostDTO) throws NotFoundException {
 
-        blogPost.setContent(content);
-        blogPost.setTimeStamp(date);
+        Long userId = userService.getAuthUser().getId();
+        User user = userRepository.findById(userId).orElseThrow();
 
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
-        System.out.println(blogPost);
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
-        return blogPostRepository.save(blogPost);
-    }
-
-    public BlogPost saveBlogPost(String content) {
         BlogPost blogPost = new BlogPost();
-        blogPost.setContent(content);
+        blogPost.setContent(blogPostDTO.getContent());
         blogPost.setTimeStamp(new Date());
-        // blogPost.setUser(userService.get);
+        blogPost.setUser(user);
 
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
-        System.out.println(blogPost);
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
-        return blogPostRepository.save(blogPost);
+        blogPost = blogPostRepository.save(blogPost);
+        return this.mapToDto(blogPost);
     }
 
-    public BlogPost deleteBlogPost(Long id) {
-        BlogPost blogPost = blogPostRepository.findById(id).orElseThrow();
+    @Transactional
+    public BlogPostDTO updateBlogPost(BlogPostDTO blogPostDTO) throws AccessDeniedException {
+        BlogPost blogPost = blogPostRepository.findById(blogPostDTO.getId())
+                .orElseThrow(() -> new NotFoundException("Blog post not found"));
+        if (userService.getAuthUser().getId() != blogPost.getUser().getId())
+            throw new AccessDeniedException("You do not have permission to perform this action");
+        blogPost.setContent(blogPostDTO.getContent());
+        blogPost.setTimeStamp(new Date());
+        blogPost = blogPostRepository.save(blogPost);
+        return this.mapToDto(blogPost);
+    }
 
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
-        System.out.println(blogPost);
-        System.out.println("\n\n--------------------------------------------------------------\n\n");
+    @Transactional
+    public BlogPostDTO deleteBlogPost(BlogPostDTO blogPostDTO) throws AccessDeniedException {
+        BlogPost blogPost = blogPostRepository.findById(blogPostDTO.getId())
+                .orElseThrow(() -> new NotFoundException("Blog post not found"));
+        if (userService.getAuthUser().getId() != blogPost.getUser().getId())
+            throw new AccessDeniedException("You do not have permission to perform this action");
 
         blogPostRepository.delete(blogPost);
-        return blogPost;
+        return this.mapToDto(blogPost);
+    }
+
+    private BlogPostDTO mapToDto(BlogPost blogPost) {
+        BlogPostDTO blogPostDTO = new BlogPostDTO(
+                blogPost.getId(),
+                userMapper.mapToDto(blogPost.getUser()), blogPost.getTimeStamp(), blogPost.getContent());
+        return blogPostDTO;
     }
 
 }
